Split diagnostics into separate module
